Move and modularise theme setup state (#43633)
Also extracts a few inline selectors to a proper selector module.
