+++ client/my-sites/site-settings/theme-setup/index.jsx
@@ -1,7 +1,6 @@
 /**
  * External dependencies
  */
-
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { localize } from 'i18n-calypso';
@@ -19,7 +18,7 @@ import ThemeSetupPlaceholder from './theme-setup-placeholder';
 import { getSelectedSiteId, getSelectedSiteSlug } from 'state/ui/selectors';
 import { getActiveTheme, getTheme } from 'state/themes/selectors';
 import { isJetpackSite } from 'state/sites/selectors';
-import { toggleDialog } from 'state/ui/theme-setup/actions';
+import { toggleDialog } from 'state/theme-setup/actions';
 
 /**
  * Style dependencies
