+++ client/state/theme-setup/actions.js
@@ -1,13 +1,14 @@
 /**
  * Internal dependencies
  */
-
 import wpcom from 'lib/wp';
 import {
 	THEME_SETUP_TOGGLE_DIALOG,
 	THEME_SETUP_REQUEST,
 	THEME_SETUP_RESULT,
 } from 'state/themes/action-types';
+
+import 'state/theme-setup/init';
 
 export function toggleDialog() {
 	return {
