+++ client/state/ui/reducer.js
@@ -20,7 +20,6 @@ import mediaModal from './media-modal/reducer';
 import postTypeList from './post-type-list/reducer';
 import preview from './preview/reducer';
 import section from './section/reducer';
-import themeSetup from './theme-setup/reducers';
 
 /**
  * Tracks the currently selected site ID.
@@ -107,7 +106,6 @@ const reducer = combineReducers( {
 	section,
 	selectedSiteId,
 	siteSelectionInitialized,
-	themeSetup,
 } );
 
 export default reducer;
